+++ interfaces/xrtm_int_idl.c
@@ -750,7 +750,7 @@ void IDL_CDECL xrtm_get_kernel_dlm(int argc, IDL_VPTR argv[], char *argk)
      else
           i_kernel = argv[1]->value.i;
      r = xrtm_get_kernel(d, i_kernel);
-     if ((int) r == XRTM_INT_ERROR)
+     if (r == XRTM_INT_ERROR)
           IDL_Message(IDL_M_NAMED_GENERIC, IDL_MSG_LONGJMP, "ERROR: xrtm_get_kernel()");
      var.type = IDL_TYP_LONG;
      var.flags = 0;
